Rename modules for better consistency with package identity
- Renamed sampler_free_sampling.py → coherent_processing.py (reflects main workflow)
- Renamed sampler_free_utils.py → utils.py (standard naming, avoids sampler_free prefix)
- Updated all import statements across the codebase to use new module names
- Updated JSON config files to reference new module names
- Preserved git history using git mv
- No functionality changed, improved naming consistency
